Get familiar with JavaScript's fundamental syntax and create a simple program that utilizes variables, conditions, loops, and functions.
Welcome to the carnival gift shop! You’ve won many tickets at the carnival; it’s time to spend them! In this project, you will write the program to show the list of gifts, the tickets you have, and buy the gift you want. In addition, you will improve your JavaScript skills and get more comfortable with it!